+++ DeusServer/HealthTimeLineComponent.h
@@ -11,8 +11,8 @@ namespace DeusServer
 
 	protected:
 		virtual void OnStart() override;
-		virtual std::shared_ptr<int> Interpolate(const int & beforeValue, uint64_t beforeValueTimestamp, const int & afterValue, uint64_t afterValueTimestamp, uint64_t currentMs) const override;
-		virtual std::shared_ptr<int> Extrapolate(const int & beforeValue, uint64_t beforeValueTimestamp, uint64_t currentMs) const override;
+		virtual std::shared_ptr<int> Interpolate(const int & beforeValue, uint32_t beforeValueTimestamp, const int & afterValue, uint32_t afterValueTimestamp, uint32_t currentMs) const override;
+		virtual std::shared_ptr<int> Extrapolate(const int & beforeValue, uint32_t beforeValueTimestamp, uint32_t currentMs) const override;
 	};
 }
