+++ DeusServer/PositionTimeLineComponent.h
@@ -29,8 +29,8 @@ namespace DeusServer
 		~PositionTimeLineComponent();
 	protected:
 		virtual void OnStart() override;
-		virtual std::shared_ptr<DeusCore::DeusVector2> Interpolate(const DeusCore::DeusVector2& beforeValue, uint64_t beforeValueTimestamp, const DeusCore::DeusVector2& afterValue, uint64_t afterValueTimestamp, uint64_t currentMs) const override;
-		virtual std::shared_ptr<DeusCore::DeusVector2> Extrapolate(const DeusCore::DeusVector2& beforeValue, uint64_t beforeValueTimestamp, uint64_t currentMs) const override;
+		virtual std::shared_ptr<DeusCore::DeusVector2> Interpolate(const DeusCore::DeusVector2& beforeValue, uint32_t beforeValueTimestamp, const DeusCore::DeusVector2& afterValue, uint32_t afterValueTimestamp, uint32_t currentMs) const override;
+		virtual std::shared_ptr<DeusCore::DeusVector2> Extrapolate(const DeusCore::DeusVector2& beforeValue, uint32_t beforeValueTimestamp, uint32_t currentMs) const override;
 
 	};
 }
